refactor[Auth]: Move login request to a separate hook
