+++ src/components/Login/Login.tsx
@@ -2,7 +2,6 @@ import { SubmitHandler, useForm } from "react-hook-form";
 import { Navigate } from "react-router-dom";
 
 import { zodResolver } from "@hookform/resolvers/zod";
-import { useMutation } from "@tanstack/react-query";
 import { z } from "zod";
 
 import Alert from "@mui/material/Alert";
@@ -11,15 +10,13 @@ import Button from "@mui/material/Button";
 import Container from "@mui/material/Container";
 import CssBaseline from "@mui/material/CssBaseline";
 
-import { postLoginData } from "../../api";
 import { AUTH_CHECK } from "../../config";
 import { useAuth } from "../../contexts/AuthContext";
+import usePostLogin from "../../hooks/usePostLogin";
 import Input from "../common/Input";
 import PasswordInput from "../common/PasswordInput";
 
 import styles from "./login.module.scss";
-
-import type { Actions, Resources } from "../../types";
 
 export interface LoginInput {
   email: string;
@@ -32,7 +29,7 @@ const loginRules = z.object({
     .trim()
     .min(1, { message: AUTH_CHECK.EMAIL.EMPTY })
     .email({ message: AUTH_CHECK.EMAIL.INVALID }),
-  password: z.string().min(8, { message: AUTH_CHECK.PASSWORD.LENGTH }),
+  password: z.string().trim().min(1, { message: AUTH_CHECK.PASSWORD.EMPTY }),
 });
 
 function Login() {
@@ -50,30 +47,21 @@ function Login() {
     resolver: zodResolver(loginRules),
   });
 
-  const { accessToken, setAccessToken, setPermissions, setUsername } =
-    useAuth();
+  const { accessToken, setAccessToken, setUsername } = useAuth();
 
-  const login = useMutation({
-    mutationFn: postLoginData,
-    onError: (error) =>
-      setError("root.serverError", { type: "401", message: error.message }),
-    onSuccess: (data) => {
-      setAccessToken(data.access_token);
-      setPermissions(
-        data.permissions.reduce(
-          (permissions, current) => {
-            permissions[current.resource] = current.actions;
-            return permissions;
-          },
-          {} as Record<Resources, Actions[]>
-        )
-      );
-      setUsername(data.name);
-    },
-  });
+  const { data, mutate, error: loginError } = usePostLogin();
+
+  if (data) {
+    setAccessToken(data.access_token);
+    setUsername(data.name);
+  }
 
   if (accessToken) {
     return <Navigate to="/home/profile" replace={true} />;
+  }
+
+  if (loginError) {
+    setError("root.serverError", { type: "401", message: loginError.message });
   }
 
   const onSubmit: SubmitHandler<LoginInput> = (data) => {
@@ -83,7 +71,7 @@ function Login() {
     payload.append("username", data.email);
     payload.append("password", data.password);
 
-    login.mutate(payload);
+    mutate(payload);
   };
 
   return (
